refactor(backend): lift weather contracts into application layer (item 3)
Remediation item 3 (DIP). The weather data contracts (WeatherConfig,
WeatherDataPoint, etc.) move to application/interfaces/weather.ts so the
IFireModelingEngine port depends on the abstraction, not infrastructure.
infrastructure/weather/types.ts becomes a re-export shim (infra -> application,
the correct inward direction) so existing importers are untouched.

Type-only, behavior-preserving: build green; full suite 420 passed. The port
no longer imports from infrastructure/.
